requirements.txt
bcrypt==3.1.6
cffi==1.12.2
Click==7.0
fastapi==0.15.0
h11==0.8.1
httptools==0.0.13
passlib==1.7.1
pycparser==2.19
pydantic==0.21
PyJWT==1.7.1
python-multipart==0.0.5
python-vlc==3.0.6109
pytubetemp==9.4.0
six==1.12.0
SQLAlchemy==1.3.3
starlette==0.11.1
uvicorn==0.6.1
uvloop==0.12.2
websockets==7.0
